Unregistering events in XSelectInput() when unregistering IM filter     callbacks may be a bad idea as others may be interested in this event.     Removed the call to XSelectInput() altogether as we are in root window     anyway (Lubos Lunak). Fix size of a variable that gets assigned the value of SmartScheduleTime     (long) to long. This should help to prevent smart scheduler lockup on     64 bit systems due to overruns (Andreas Schwab).

+++ dix/dispatch.c
@@ -362,7 +362,7 @@ Dispatch(void)
     register int	nready;
     register HWEventQueuePtr* icheck = checkForInput;
 #ifdef SMART_SCHEDULE
-    int			start_tick;
+    long			start_tick;
 #endif
 
     nextFreeClientID = 1;
